Fix build for 2.6.0 release branch
Per the readme of `babel-plugin-feature-flags`: "This plugin is for Babel 6. If you need to support Babel 5 use the 0.2.x releases.".

The release versions of `babel-plugin-feature-flags` seem to be in a messed up state. 0.2.1 release contains a [fix](https://github.com/ember-cli/babel-plugin-feature-flags/commit/b120f9f905c42372cf661f8a2e5932b103b7a9d5) which ports to Babel 6 and should have gone in 0.3.0 release. This causes the build of 2.6.0 to fail.

This PR locks the version of the plugin to 0.2.0.

PS: I did not find a `release-2-6` branch and figured 2.6 was released from this branch. So I did a diff against this branch. Build on master seems to work with Babel 6 so probably this needs to be ported to previous release branches too.
(cherry picked from commit dd026832597ecad59116c76f2f6de1955ff3c150)